Separate SDL low-level and java high-level transformation of schema
